Clojure: Give up on how to refactor the stupid sums
Was trying to do this, but I can't get the `include-any?` to work
correctly. It won't let me have nested lambdas, and I can't use the let
to name the lambda either, because it gets its two parameters from two
different lambdas, so to pass one to the other I'll ultimately need to
wrap it in another lambda since it won't curry: `(#(= %1 %2) 1)`

```
(defn any? [f coll]
  (reduce #(or %1 (f %2)) false coll))

(defn include-any? [haystack & needles]
  (let [hay-is-needle? #(= %1 %2)]
       (any? #(some #(hay-is-needle? %) haystack) needles)))
```