JavaScript: move more logic into the copy function
You could rightly criticize this move, because it couples the function
to the one use case, but I think it makes more sense since this program
is so small and it simplifies the call site. As such a program grew in
complexity, it would very possibly make sense to rever this, but for now,
it's easier to understand and reduces duplication.