Rust: Jesus Christ, I know what I want to say and cannot fucking figure out how to say it.
I want to say: I'll borrow the fucking vector of borrowed fucking strings, and I won't deallocate any of them,
and I won't mutate the fucking vector or any of the fucking strings, I will iterate over the fucking vector,
and I will return to you a vector with the same lifetime as the fucking vector that you gave me, maybe it's
the same one, maybe it's a new one, whatever, just free them at the same time. That returned vector may
be filled with strings that I pulled from stdin, so I assume I need to take lifetime ownership over those strings
and set it to the lifetime of the fucking borrowed strings I received, because I want to put them into the
fucking vector I'm returning.

Or, in Ruby:

```sh
$ printf "stdin1\nstdin2" | ruby -e 'def get_paths(args, stdin) args.any? ? args : $stdin.readlines end; p get_paths(ARGV, $stdin)'
["stdin1\n", "stdin2"]

$ printf "stdin1\nstdin2" | ruby -e 'def get_paths(args, stdin) args.any? ? args : $stdin.readlines end; p get_paths(ARGV, $stdin)' argv1 argv2
["argv1", "argv2"]
```