Elixir: Uhm, it was just broken for some reason (maybe I upgraded my Elixir?)
I had to fuck with it for a bit to get it working again, it seems stupider.
Eg I set a variable, then check what its set to, if I dislike it, I reset it.
But the compiler isn't able to recognize that my resetting it doesn't mean that
I never set it in the first place, so it just whines at me and I have to add an else clause.
Also, it just blew up when starting. I removed the logger thing that mix generated,
that seemed to improve it. Also for some reason, empty paths were getting through.
Not sure why, I should probably look into that. But I'm not going to.