Fortran: Use command_argument_count over iargc and get_command_argument over getarg
Based on VladimirF's comment in http://stackoverflow.com/questions/39812704/how-to-invoke-a-program-and-pass-it-standard-input
Couldn't use the env var, b/c the env var is set by the shell, not the OS, so when the test program changes directory
and then invokes the binary, there is nothing updating the environment variable. This means that the program gets confused
about where the user is located at on the file system, and prints the wrong path.