Lua: refactor `getEach` to delegate to `ipairs`
Mostly b/c I don't understand why it receives the table twice (once to `itable`
and then again on each invocation to the getter that it returns). So I'm
hesitant to reimplement it. Also b/c I don't know the cost of doing `#t`
on each iteration through the loop. And b/c it's apparently optimized for
tables that are really arrays (sequential integer keys), so I figure it's best
to just delegate to it and let it do its own thing